introduced a socket_type to hide linux and windows type differences

+++ include/matador/net/socket_stream.hpp
@@ -80,14 +80,14 @@ template < class P >
 template < class Buffer >
 ssize_t socket_stream<P>::receive(Buffer &buffer)
 {
-  return ::recv(this->id(), buffer.data(), buffer.capacity(), 0);
+  return ::recv(this->id(), buffer.data(), static_cast<int>(buffer.capacity()), 0);
 }
 
 template < class P >
 template < class Buffer >
 ssize_t socket_stream<P>::send(const Buffer &buffer)
 {
-  return ::send(this->id(), buffer.data(), buffer.size(), 0);
+  return ::send(this->id(), buffer.data(), static_cast<int>(buffer.size()), 0);
 }
 /// @endcond
 }
